bug(server): fix build script - restructures application folder structure - rewrites build script using the correct directory [Finishes #166983634]
